update profile fix
