+++ apps/account/api_endpoints/auth/__init__.py
@@ -3,3 +3,4 @@ from .Register import *
 from .profile import *
 from .change_password import *
 from .forgot_password import *
+from .update_profile import *
