+++ apps/account/api_endpoints/auth/update_profile/__init__.py
@@ -1,0 +1,1 @@
+from .views import *
