+++ apps/account/api_endpoints/auth/update_profile/serializers.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+
+from apps.account.models import Users
+
+
+class ProfileUpdateSerializers(serializers.ModelSerializer):
+
+    class Meta:
+        model = Users
+        fields = ['first_name', 'last_name', 'phone_number']
+
